refactor: separate features into diff modules
