--- docs/python_api/datasets/l2g_feature.md
@@ -1,29 +1,0 @@
----
-title: L2G Feature
----
-
-## Abstract Class
-
-::: gentropy.dataset.l2g_feature.L2GFeature
-
-## Feature Classes
-
-### Derived from colocalisation
-
-::: gentropy.dataset.l2g_feature.EQtlColocClppMaximumFeature
-::: gentropy.dataset.l2g_feature.PQtlColocClppMaximumFeature
-::: gentropy.dataset.l2g_feature.SQtlColocClppMaximumFeature
-::: gentropy.dataset.l2g_feature.TuQtlColocClppMaximumFeature
-::: gentropy.dataset.l2g_feature.EQtlColocH4MaximumFeature
-::: gentropy.dataset.l2g_feature.PQtlColocH4MaximumFeature
-::: gentropy.dataset.l2g_feature.SQtlColocH4MaximumFeature
-::: gentropy.dataset.l2g_feature.TuQtlColocH4MaximumFeature
-
-### Derived from distance
-
-::: gentropy.dataset.l2g_feature.DistanceTssMinimumFeature
-::: gentropy.dataset.l2g_feature.DistanceTssMeanFeature
-
-## Schema
-
---8<-- "assets/schemas/l2g_feature.md"
