+++ docs/python_api/datasets/l2g_features/_l2g_feature.md
@@ -1,0 +1,11 @@
+---
+title: L2G Feature
+---
+
+## Abstract Class
+
+::: gentropy.dataset.l2g_features.L2GFeature
+
+## Schema
+
+--8<-- "assets/schemas/l2g_feature.md"
