+++ docs/python_api/datasets/l2g_features/distance.md
@@ -1,0 +1,19 @@
+---
+title: Distance derived features
+---
+
+## List of features
+
+::: gentropy.dataset.l2g_features.DistanceTssMinimumFeature
+::: gentropy.dataset.l2g_features.DistanceTssMeanFeature
+::: gentropy.dataset.l2g_features.DistanceFootprintMinimumFeature
+::: gentropy.dataset.l2g_features.DistanceFootprintMeanFeature
+::: gentropy.dataset.l2g_features.DistanceTssMinimumNeighbourhoodFeature
+::: gentropy.dataset.l2g_features.DistanceTssMeanNeighbourhoodFeature
+::: gentropy.dataset.l2g_features.DistanceFootprintMinimumNeighbourhoodFeature
+::: gentropy.dataset.l2g_features.DistanceFootprintMeanNeighbourhoodFeature
+
+## Common logic
+
+::: gentropy.dataset.l2g_features.\_common_distance_feature_logic
+::: gentropy.dataset.l2g_features.\_common_neighbourhood_distance_feature_logic
