+++ src/gentropy/dataset/l2g_features/__init__.py
@@ -1,0 +1,3 @@
+"""Feature factories for L2G."""
+
+from __future__ import annotations
